wip(mir): ADR-040 sayısal tip genişlikleri — longint/float32/double (#113)
int=32-bit (tanımlı wrap, önceki commit'te kapandı), longint=64-bit (rank
kulesi dışı izole tip), float=gerçek 32-bit IEEE single (ayrı ValueKind,
önceden yanlışlıkla double'a alias'lanıyordu), double=64-bit IEEE, byte
aritmetikte int'e terfi eden aralık-kısıtlı tip (gerçek 8-bit wrap DEĞİL).
Kapsar: Type/Value/IR opcode şeması (LOAD_LONG/LADD.../LOAD_FLOAT32/
F32ADD... + CAST_LONG_TO_STR/CAST_FLOAT32_TO_STR vb.), IRGenerator slot tip
çözümü ve yayılımı, MIR JIT backend'de longint (MIR_T_I64) ve float32
(MIR_T_F, gerçek single — F2D/D2F trampolinleri) codegen + rt_jit_* runtime
fonksiyonları, constant folding/DAP/host_functions/root.sqt FFI imza
güncellemeleri, yeni golden testler (tests/golden/numeric/widths,
longint_narrowing_compile_error) + tests/golden/float/basic.sqt'nin
float→double dönüşümü (ADR-040: çıplak ondalık literal artık double
varsayılan; basic.expected İÇERİĞİ DEĞİŞMEDİ — float32/double bu testteki
basit değerlerde aynı string temsilini üretiyor).

DURUM: TAMAMLANMAMIŞ / test kırmızı — kesinlikle "Faz 1-4 tamam" olarak
işaretlenmemeli. İzole build+test doğrulaması:
  ninja -C build → temiz derleniyor (yalnızca -Wswitch uyarıları: VM'in
  birkaç yardımcı switch'i — runUntilEvent GC-adım tespiti, valueEqual,
  valueToJsonStr — yeni LongInt/Float32 opcode/ValueKind dallarını henüz
  ele almıyor; çalışma zamanı hatasına yol açmıyor ama sessiz atlama riski
  taşıyor, ayrı bir borç).
  bash tests/run.sh → 65 geçti, 4 başarısız; ctest → 110/116 yeşil.
  Başarısızlar:
  - tests/golden/numeric/widths: DERLENMİYOR BİLE — longint hedefli
    değişkene int/longint literal ataması "cannot assign int to longint"
    ile reddediliyor VE longint aritmetiği "arithmetic operator on
    non-numeric type: longint and int/longint" ile reddediliyor;
    type_checker.cpp'deki ADR-040 checkAssign/checkExpr özel-durum
    dalları (longint'i lossless kabul etmesi gereken kod) beklendiği
    gibi tetiklenmiyor — kök neden araştırılmadı.
  - tests/golden/numeric/longint_narrowing_compile_error: beklenen TEK
    hata (`longint→int requires explicit cast`) yerine biri SPÜRIYÖZ
    olmak üzere 2 hata üretiyor (`longint L = 5000000000;` — hedef zaten
    longint olduğu halde "cannot assign int to longint" ile reddediliyor;
    bu satırın derlenmesi gerekiyordu).
  - tests/golden/math/pi_e, tests/golden/math/ffi_import,
    tests/golden/sys/random_range: sayısal değer/karşılaştırma farklılığı
    (beklenen "11"/"53.5..." yerine "00"/"50.0..." vb.) — muhtemelen
    float32↔double karşılaştırma/format değişikliğiyle ilgili, kök neden
    araştırılmadı.

Bu commit'in içeriği tamamen çalışma ağacında ZATEN mevcut olan (bu görev
başlamadan önceki) WIP'in birebir aktarımıdır — hiçbir yeni düzeltme
denemesi yapılmadı (kapsam: commit'lere ayırma, ADR-040'ı bitirmek değil).

+++ src/ffi/root_sqt.hpp
@@ -24,22 +24,24 @@ inline const char* kEmbeddedRootSqt = R"SQT(
 ffi string version() : CORE_VERSION from core;   // derleyici sürümü (0.8.0)
 
 // ── math modülü (saf hesap, capability'siz — #89) ──────────────────────────
-// Overload yok → int/float ayrımı isimle. sqrt(-1)=NaN, Error fırlatmaz.
-ffi int   abs(int x)              : MATH_ABS   from math;
-ffi float absf(float x)           : MATH_ABSF  from math;
-ffi int   min(int a, int b)       : MATH_MIN   from math;
-ffi int   max(int a, int b)       : MATH_MAX   from math;
-ffi float minf(float a, float b)  : MATH_MINF  from math;
-ffi float maxf(float a, float b)  : MATH_MAXF  from math;
-ffi float sqrt(float x)           : MATH_SQRT  from math;
-ffi float pow(float b, float e)   : MATH_POW   from math;
-ffi float floor(float x)          : MATH_FLOOR from math;
-ffi float ceil(float x)           : MATH_CEIL  from math;
-ffi float round(float x)          : MATH_ROUND from math;
+// Overload yok → int/ondalık ayrımı isimle (abs/absf). ADR-040: ondalık
+// matematik 64-bit `double` üzerinden (isimlerdeki "f" tarihsel; tip double).
+// sqrt(-1)=NaN, Error fırlatmaz.
+ffi int    abs(int x)               : MATH_ABS   from math;
+ffi double absf(double x)           : MATH_ABSF  from math;
+ffi int    min(int a, int b)        : MATH_MIN   from math;
+ffi int    max(int a, int b)        : MATH_MAX   from math;
+ffi double minf(double a, double b) : MATH_MINF  from math;
+ffi double maxf(double a, double b) : MATH_MAXF  from math;
+ffi double sqrt(double x)           : MATH_SQRT  from math;
+ffi double pow(double b, double e)  : MATH_POW   from math;
+ffi double floor(double x)          : MATH_FLOOR from math;
+ffi double ceil(double x)           : MATH_CEIL  from math;
+ffi double round(double x)          : MATH_ROUND from math;
 // #89: gerçek importable sabit yok (ffi yalnızca fonksiyon) — sıfır-argümanlı
 // saf fonksiyon olarak sunulur: import {PI} from math; PI();
-ffi float PI()                    : MATH_PI    from math;
-ffi float E()                     : MATH_E     from math;
+ffi double PI()                     : MATH_PI    from math;
+ffi double E()                      : MATH_E     from math;
 
 // ── caps modülü (pledge modeli, capability'siz — #91) ───────────────────────
 // drop geri alınamaz; ekleme fonksiyonu YOK (güvenlik değeri buradan gelir).
